docker/requirements.txt
boto3==1.24.34
pandas==1.4.3
plotly==5.9.0
requests==2.28.1
streamlit==1.11.0
streamlit-aggrid==0.2.3.post2
pyyaml